requirements.txt
fastapi
uvicorn
python-dotenv
passlib[bcrypt]
python-jose
faiss-cpu
sentence-transformers
rank-bm25
pymupdf
python-docx
python-pptx
pandas
beautifulsoup4
requests
pytesseract
Pillow
streamlit
groq